Moves transaction generators to seperate folders
