+++ app/modules/tx/transactionsGenerators/withdraw/sagas.ts
@@ -1,0 +1,44 @@
+import { BigNumber } from "bignumber.js";
+import { put, select } from "redux-saga/effects";
+
+import { TGlobalDependencies } from "../../../../di/setupBindings";
+import { ITxData } from "../../../../lib/web3/Web3Manager";
+import { IAppState } from "../../../../store";
+import { actions, TAction } from "../../../actions";
+import {
+  selectEtherTokenBalance,
+} from "../../../wallet/selectors";
+import { selectEthereumAddressWithChecksum } from "../../../web3/selectors";
+
+export function* generateEthWithdrawTransaction(
+  { contractsService }: TGlobalDependencies,
+  action: TAction,
+): any {
+  if (action.type !== "TX_SENDER_ACCEPT_DRAFT") return;
+
+  const txStateDetails = action.payload;
+
+  let txDetails: ITxData | undefined;
+  if (!txStateDetails) return;
+  const etherTokenBalance = yield select((s: IAppState) => selectEtherTokenBalance(s.wallet));
+  const from = yield select((s: IAppState) => selectEthereumAddressWithChecksum(s.web3));
+
+  // transaction can be fully covered by etherTokens
+
+  const txInput = contractsService.etherToken
+    .withdrawAndSendTx(txStateDetails.to, new BigNumber(txStateDetails.value))
+    .getData();
+  const ethVal = new BigNumber(txStateDetails.value);
+  const difference = ethVal.sub(etherTokenBalance);
+
+  txDetails = {
+    to: contractsService.etherToken.address,
+    from,
+    data: txInput,
+    value: difference.comparedTo(0) > 0 ? difference.toString() : "0",
+    gas: txStateDetails.gas,
+    gasPrice: txStateDetails.gasPrice,
+  };
+
+  yield put(actions.txSender.txSenderAcceptDraft(txDetails));
+}
